moved components to respective space, and removed components from useVolunteerSteps and made it common
