+++ view/src/components/common/StaggerItem.tsx
@@ -1,0 +1,29 @@
+import { motion } from 'framer-motion';
+
+const fadeUpSpring = {
+  hidden: { opacity: 0, y: 16 },
+  show: {
+    opacity: 1,
+    y: 0,
+    transition: {
+      type: 'spring',
+      stiffness: 420,
+      damping: 32,
+      mass: 0.8,
+    },
+  },
+};
+
+export function StaggerItem({
+  children,
+  className,
+}: {
+  children: React.ReactNode;
+  className?: string;
+}) {
+  return (
+    <motion.div variants={fadeUpSpring} className={className}>
+      {children}
+    </motion.div>
+  );
+}
